Add a tool to convert MediaWiki markdown to hugo.
I wrote this tool to convert content of a site I maintain.

In my case, I did not have a MediaWiki installation anymore, but an XML
file and a filesystem backup.

I used https://github.com/outofcontrol/mediawiki-to-gfm to convert the
XML to Markdown files, and then I needed to clean up the Markdown to
that the site is accepted by Hugo - no dangling references, links work,
broken links are removed, etc.